pyproject.toml
[tool.poetry]
name = "bot-discord-ia-lastwar"
version = "0.1.0"
description = ""
authors = ["Calel <borgescalel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mkdocs = "^1.6.1"
pydantic = "^2.11.7"
pytest = "^8.4.1"
mkdocs-material = "^9.6.15"
python-dotenv = "^1.1.1"
dotenv = "^0.9.9"
taskipy = "^1.14.1"
sentry-sdk = "^2.32.0"
discord-py = "^2.5.2"
bs4 = "^0.0.2"
google-genai = "^1.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
test = "lsof -ti :8501 | xargs kill -9 | pytest tests -v"
